- fixed something in calculate_square_rows_cols

+++ vision/utils/plot_utils.py
@@ -108,5 +108,5 @@ def noticks():
 
 def calculate_square_rows_cols(n):
     rows = int(np.ceil(np.sqrt(n)))
-    cols = np.ceil(n/rows)
+    cols = int(np.ceil(n/rows))
     return rows, cols
